Update API-design-guidelines.md
Following May 4th Commonalities, we agreed to update the Notification chapter with following points:
- Extend `notificationAuthToken` to 256 as max length
- Explicitly mention OAuth 2.0 for `notificationAuthToken`
- add `correlationId` management in subscriptions and notification
- rename `subscriptions' to `eventSubscriptions` - for consistency I took the initiative to align in same way `notifications` to `eventNotifications`

Ready for review.
As stated during the meeting:  we have to move forward on this topic as API projects are waiting for this. So please, provide in this PR only comment for blocking issue before first release. 
Ideas & suggestions for improvement are welcome but should be logged in independent issue.