Create APIProposal_E164_E212_MAP_TelecomsXChange.md
The Dial Code Lookup API provides up-to-date information on the mapping of dial codes (also known as calling codes or country codes) to countries and regions. Currently, there is no standardized mapping list for dial codes, and each operator creates their own list, which is often out-of-date. This API aims to standardize the mapping list to increase the accuracy of billing and routing and reduce disputes related to mis-matched dial code mapping. By accessing the Camara API, telecom solution software developers, such as IMS, Softswitches, and SMSCs, can easily and automatically pull updated dial codes as they change, improving the accuracy and reliability of their systems.